requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
blinker==1.9.0
    # via flask
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
flask==3.1.0
    # via -r requirements.in
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
pydantic==2.10.5
    # via -r requirements.in
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via flask
